backend/requirements.txt
bidict==0.22.1
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
python-engineio==4.3.4
python-socketio==5.7.2
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
